+++ changelog.md
@@ -1,5 +1,28 @@
 # Changelog
 ## Week 1
-### 05/02/2024
 - Tabel met teams en beschrijving toegevoegd
 - Overzicht pagina basic styling gegeven
+
+- API onderzoek gedaan, oefening gedaan om het beter te begrijpen
+Resources: https://snipcart.com/blog/topics/web-development
+https://snipcart.com/blog/integrating-apis-introduction
+
+document.addEventListener("click", function (event) {
+  // Checking if the button was clicked
+  if (!event.target.matches("#button")) return;
+
+  console.log("Button was clicked!");
+});
+
+- downloaded github app
+
+### Javascript
+
+- JSON bestand toegevoegd met de hulp van manoah
+- 
+- GetElementById
+
+[! javascript stats](/images/stats.png)
+[! html](/images/html.png)
+
+